L1prio que

+++ SOCKET.go
@@ -25,7 +25,9 @@ var (
 )
 
 func (his *HISTORY) startSocket(tcpListen string) {
-
+	if BootHisCli {
+		return
+	}
 	// socket listener
 	go func() {
 		os.Remove(SocketPath)
@@ -171,8 +173,7 @@ forever:
 				case CaseDupes:
 					// we locked the hash but IndexQuery replied with Duplicate
 					// set L1 cache to Dupe and expire
-					//his.L1Cache.Set(hobj.MessageIDHash, hobj.Char, CaseDupes, FlagExpires)
-					//his.DoCacheEvict(hobj.Char, &hobj.MessageIDHash, 0, &EmptyStr)
+					his.DoCacheEvict(hobj.Char, hobj.MessageIDHash, 0, EmptyStr)
 					//dupes++
 					if hobj.ResponseChan != nil {
 						hobj.ResponseChan <- isDup
@@ -181,8 +182,7 @@ forever:
 				case CaseRetry:
 					// we locked the hash but IndexQuery replied with Retry
 					// set L1 cache to Retry and expire
-					//his.L1Cache.Set(hobj.MessageIDHash, hobj.Char, CaseRetry, FlagExpires)
-					//his.DoCacheEvict(hobj.Char, &hobj.MessageIDHash, 0, &EmptyStr)
+					his.DoCacheEvict(hobj.Char, hobj.MessageIDHash, 0, EmptyStr)
 					//retry++
 					if hobj.ResponseChan != nil {
 						hobj.ResponseChan <- isDup
